backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
bcrypt==5.0.0
billiard==4.2.2
blinker==1.9.0
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
dnspython==2.8.0
Flask==3.0.3
flask-cors==6.0.2
Flask-WTF==1.2.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.3
marshmallow==3.21.3
mongo==0.2.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.11.0
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==7.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.5
WTForms==3.2.1
